KS-31 fix: 회원 존재 유무 및 등록한 회원 확인 로직 추가

+++ src/main/java/com/kreamish/kream/sale/service/SaleServiceImpl.java
@@ -112,13 +112,20 @@ public class SaleServiceImpl implements SaleService {
 
     @Override
     public SaleDeleteResponseDto withdrawSale(Long memberId, Long saleId) {
+        Member member = memberRepository.findById(memberId)
+            .orElseThrow(() -> new NoSuchElementException("Not Found Member By MemberId"));
+
         Sale sale = saleRepository.findById(saleId)
             .orElseThrow(() -> new NoSuchElementException("Not Found Sale by saleId"));
+
+        if(!sale.getMember().getMemberId().equals(member.getMemberId())){
+            throw new IllegalStateException("Sale is not registered by current member");
+        }
+        dealStatusShouldPending(sale.getSaleStatus());
 
         Long itemSizesId = sale.getItemSizes().getItemSizesId();
         Long beforePrice = sale.getSalePrice();
 
-        dealStatusShouldPending(sale.getSaleStatus());
 
         saleRepository.deleteById(sale.getSaleId());
